Removed tables_plus and more changes for tableaux

+++ notes/views.py
@@ -1,45 +1,64 @@
-from django.shortcuts import render
+from django.shortcuts import render, reverse
 from django.contrib.auth.mixins import LoginRequiredMixin
 from notes.models import Note
 from notes.forms import NoteForm
 from notes.tables import NoteTable
 from django.views.generic import CreateView, UpdateView, DetailView, View
 from table_manager.views import FilteredTableView, AjaxCrudView
+from django_tableaux.views import TableauxView, SelectedMixin
+from django_tableaux.buttons import Button
 from shop.models import Item
+from django_htmx.http import HttpResponseClientRefresh
 
 
-class NoteListView(LoginRequiredMixin, FilteredTableView):
+class NoteListView(LoginRequiredMixin, TableauxView):
     model = Note
     table_class = NoteTable
     table_pagination = {"per_page": 100}
-    allow_create = True
-    allow_update = True
+    template_name = "shop/table_wide.html"
+    column_settings = True
+    click_action = TableauxView.ClickAction.HX_GET
+    click_url_name = "notes:update"
 
     def get_queryset(self):
         return Note.objects.all().order_by("title")
 
+    def get_actions(self):
+        return (("delete", "Delete"),)
 
-class NoteCreateView(LoginRequiredMixin, AjaxCrudView):
-    model = Note
-    form_class = NoteForm
+    def handle_action(self, request, action):
+        if action == "delete":
+            self.selected_objects.delete()
 
-    def save_object(self, data, **kwargs):
-        super().save_object(data, **kwargs)
-        self.object.user = self.request.user
-        self.object.save()
-
-
-class NoteUpdateView(NoteCreateView):
-    update = True
-    allow_delete = True
+    def get_buttons(self):
+        return [
+            Button("New note", hx_get=reverse("notes:htmx"), hx_target="#modals-here")
+        ]
 
 
 class NoteHtmxView(LoginRequiredMixin, View):
+    """
+    This view can be called from an item page or from the notes list page
+    It handles both update and creation
+    """
+
     def get(self, request, *args, **kwargs):
-        bits = request.htmx.trigger_name.split("-")
-        item = Item.objects.filter(pk=bits[1]).first()
-        note = Note.objects.filter(item=item).first() if item else None
         template = "notes/note_form.html"
+        pk = kwargs.get("pk", None)
+        if pk:
+            # Called from note list view
+            note = Note.objects.filter(pk=pk).first()
+            item = note.item
+        elif request.htmx.trigger_name == "new-note":
+            # Create called from note list view
+            note = None
+            item = None
+        else:
+            # Called from item view - trigger name contains item pk
+            bits = request.htmx.trigger_name.split("-")
+            item = Item.objects.filter(pk=bits[1]).first()
+            note = Note.objects.filter(item=item).first() if item else None
+
         if note:
             form = NoteForm(initial={"title": note.title, "content": note.content})
         else:
@@ -48,11 +67,12 @@ class NoteHtmxView(LoginRequiredMixin, View):
         return render(request, template, context)
 
     def post(self, request, *args, **kwargs):
-        template = "notes/close_modal.html"
         form = NoteForm(request.POST)
-        item = Item.objects.get(pk=request.POST["item"])
-        note = Note.objects.filter(item=item).first() if item else None
-        if "delete" in request.POST:
+        item_pk = request.POST.get("item_pk", None)
+        item = Item.objects.get(pk=item_pk) if item_pk else None
+        note_pk = request.POST.get("note_pk", None)
+        note = Note.objects.filter(pk=note_pk).first() if note_pk else None
+        if note and "delete" in request.POST:
             note.delete()
             note = None
         elif form.is_valid():
@@ -63,16 +83,22 @@ class NoteHtmxView(LoginRequiredMixin, View):
                     note.delete()
                     note = None
             else:
+                changed = False
                 if not title:
-                    title = item.ref
+                    title = item.ref if item else "Untitled"
+                    changed = True
                 if note:
-                    note.title = title
-                    note.content = content
-                    note.user = request.user
-                    note.save()
+                    if note.title != title:
+                        note.title = title
+                        changed = True
+                    if note.content != content:
+                        note.content = content
+                        changed = True
+                    if changed:
+                        note.user = request.user
+                        note.save()
                 else:
                     note = Note.objects.create(
                         title=title, content=content, item=item, user=request.user
                     )
-        context = {"note": note, "item": item}
-        return render(request, template, context)
+        return HttpResponseClientRefresh()
